fix news

+++ src/main/java/com/example/webBanXeApi/models/News.java
@@ -22,7 +22,7 @@ public class News {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String title;
@@ -44,11 +44,11 @@ public class News {
 
     // getters and setters
 
-public Integer getId() {
+public Long getId() {
     return id;
 }
 
-public void setId(Integer id) {
+public void setId(long id) {
     this.id = id;
 }
